requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# WhatsApp Integration
twilio==8.11.0

# LLM & AI
openai==1.10.0
anthropic==0.18.0

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Database & Caching
redis==5.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.26.0

# Document Processing (for knowledge base)
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
